changed from float to double

+++ driver_table_reader.h
@@ -13,17 +13,17 @@
 // a data structure to hold the table information,
 // including the values of the headers
 struct Table {
-  float *data;
+  double *data;
   int stride_energy;
   int stride_log_Rs;
   int stride_log_Ts;
-  float energies[DIM_R];
-  float log_Ts[DIM_S];
-  float log_Rs[DIM_F];
+  double energies[DIM_R];
+  double log_Ts[DIM_S];
+  double log_Rs[DIM_F];
 };
 typedef struct Table Table;
 
-float *get_table_row(Table *, int, int, int);
+double *get_table_row(Table *, int, int, int);
 
 void initialise_table_from_file(Table *, char *);
